refactor: rework party lookup loop
