+++ modules/api-svc/package.json
@@ -95,7 +95,7 @@
     "@babel/core": "^7.19.1",
     "@babel/preset-env": "^7.19.1",
     "@redocly/openapi-cli": "^1.0.0-beta.94",
-    "@types/jest": "^29.0.2",
+    "@types/jest": "^29.0.3",
     "babel-jest": "^29.0.3",
     "eslint": "^8.23.1",
     "eslint-config-airbnb-base": "^15.0.0",
@@ -104,7 +104,7 @@
     "jest": "^29.0.3",
     "jest-junit": "^14.0.1",
     "nock": "^13.2.9",
-    "npm-check-updates": "^16.1.2",
+    "npm-check-updates": "^16.1.3",
     "openapi-response-validator": "^12.0.2",
     "openapi-typescript": "^5.4.1",
     "redis-mock": "^0.56.3",
