+++ modules/outbound-command-event-handler/package.json
@@ -54,7 +54,7 @@
   "devDependencies": {
     "@types/convict": "^6.1.1",
     "@types/express": "^4.17.14",
-    "@types/jest": "^29.0.2",
+    "@types/jest": "^29.0.3",
     "@types/node": "^18.7.18",
     "@types/node-cache": "^4.2.5",
     "@types/supertest": "^2.0.12",
@@ -66,7 +66,7 @@
     "eslint": "^8.23.1",
     "jest": "^29.0.3",
     "nodemon": "^2.0.19",
-    "npm-check-updates": "^16.1.2",
+    "npm-check-updates": "^16.1.3",
     "replace": "^1.2.1",
     "standard-version": "^9.5.0",
     "ts-jest": "^29.0.1",
