+++ modules/outbound-command-event-handler/test/integration/application/discovery/process_bulk_accept_party_info.test.ts
@@ -39,7 +39,6 @@ import {
   IProcessPartyInfoCallbackCmdEvtData,
   IProcessSDKOutboundBulkAcceptPartyInfoCmdEvtData,
   IProcessSDKOutboundBulkPartyInfoRequestCmdEvtData,
-  IProcessSDKOutboundBulkPartyInfoRequestCompleteCmdEvtData,
   IProcessSDKOutboundBulkRequestCmdEvtData,
   IRedisBulkTransactionStateRepoOptions,
   KafkaCommandEventProducer,
@@ -47,7 +46,6 @@ import {
   ProcessPartyInfoCallbackCmdEvt,
   ProcessSDKOutboundBulkAcceptPartyInfoCmdEvt,
   ProcessSDKOutboundBulkPartyInfoRequestCmdEvt,
-  ProcessSDKOutboundBulkPartyInfoRequestCompleteCmdEvt,
   ProcessSDKOutboundBulkRequestCmdEvt,
   RedisBulkTransactionStateRepo,
 } from '@mojaloop/sdk-scheme-adapter-private-shared-lib'
@@ -235,18 +233,6 @@ describe("Tests for ProcessSDKOutboundBulkAcceptPartyInfo Event Handler", () => 
     await producer.sendCommandEvent(processPartyInfoCallbackMessageObjOne);
     const processPartyInfoCallbackMessageObjTwo = new ProcessPartyInfoCallbackCmdEvt(processPartyInfoCallbackMessageData2);
     await producer.sendCommandEvent(processPartyInfoCallbackMessageObjTwo);
-    await new Promise(resolve => setTimeout(resolve, messageTimeout));
-
-    // Simulate the domain handler sending the command handler ProcessSDKOutboundBulkPartyInfoRequestComplete message
-    const processSDKOutboundBulkPartyInfoRequestCompleteCommandEventData : IProcessSDKOutboundBulkPartyInfoRequestCompleteCmdEvtData = {
-      bulkId: bulkTransactionId,
-      timestamp: Date.now(),
-      headers: []
-    }
-    const processSDKOutboundBulkPartyInfoRequestCompleteCommandEventObj = new ProcessSDKOutboundBulkPartyInfoRequestCompleteCmdEvt(
-      processSDKOutboundBulkPartyInfoRequestCompleteCommandEventData
-    );
-    await producer.sendCommandEvent(processSDKOutboundBulkPartyInfoRequestCompleteCommandEventObj);
     await new Promise(resolve => setTimeout(resolve, messageTimeout));
 
     // Command event for sdk outbound bulk accept party info request
