+++ modules/outbound-command-event-handler/test/integration/application/discovery/process_bulk_request.test.ts
@@ -27,32 +27,26 @@
  import { DefaultLogger } from "@mojaloop/logging-bc-client-lib";
  import { ILogger } from "@mojaloop/logging-bc-public-types-lib";
  import { SDKSchemeAdapter } from '@mojaloop/api-snippets';
- 
+
  import {
    DomainEvent,
    IKafkaEventConsumerOptions,
    IKafkaEventProducerOptions,
-   IProcessPartyInfoCallbackCmdEvtData,
-   IProcessSDKOutboundBulkPartyInfoRequestCmdEvtData,
-   IProcessSDKOutboundBulkPartyInfoRequestCompleteCmdEvtData,
    IProcessSDKOutboundBulkRequestCmdEvtData,
    IRedisBulkTransactionStateRepoOptions,
    KafkaCommandEventProducer,
    KafkaDomainEventConsumer,
-   ProcessPartyInfoCallbackCmdEvt,
-   ProcessSDKOutboundBulkPartyInfoRequestCmdEvt,
-   ProcessSDKOutboundBulkPartyInfoRequestCompleteCmdEvt,
    ProcessSDKOutboundBulkRequestCmdEvt,
    RedisBulkTransactionStateRepo,
  } from '@mojaloop/sdk-scheme-adapter-private-shared-lib'
  import { randomUUID } from "crypto";
- 
+
  // Tests can timeout in a CI pipeline so giving it leeway
  jest.setTimeout(10000)
- 
+
  const logger: ILogger = new DefaultLogger('bc', 'appName', 'appVersion'); //TODO: parameterize the names here
  const messageTimeout = 2000;
- 
+
  // Setup for Kafka Producer
  const commandEventProducerOptions: IKafkaEventProducerOptions = {
    brokerList: 'localhost:9092',
@@ -60,7 +54,7 @@
    topic: 'topic-sdk-outbound-command-events'
  }
  const producer = new KafkaCommandEventProducer(commandEventProducerOptions, logger)
- 
+
  // Setup for Kafka Consumer
  const domainEventConsumerOptions: IKafkaEventConsumerOptions = {
    brokerList: 'localhost:9092',
@@ -74,38 +68,38 @@
    domainEvents.push(message);
  }
  const consumer = new KafkaDomainEventConsumer(_messageHandler.bind(this), domainEventConsumerOptions, logger)
- 
+
  // Setup for Redis access
  const bulkTransactionEntityRepoOptions: IRedisBulkTransactionStateRepoOptions = {
    connStr: 'redis://localhost:6379'
  }
  const bulkTransactionEntityRepo = new RedisBulkTransactionStateRepo(bulkTransactionEntityRepoOptions, logger);
- 
- 
+
+
  describe("Tests for discovery part in Outbound Command Event Handler", () => {
- 
+
    beforeEach(async () => {
      domainEvents = [];
    });
- 
+
    beforeAll(async () => {
      await producer.init();
      await consumer.init();
      await consumer.start();
      await bulkTransactionEntityRepo.init();
    });
- 
+
    afterAll(async () => {
      await producer.destroy();
      await consumer.destroy();
      await bulkTransactionEntityRepo.destroy();
    });
- 
+
    // TESTS FOR PARTY LOOKUP
    test.only("1. When inbound command event ProcessSDKOutboundBulkRequest is received \
           Then outbound event SDKOutboundBulkPartyInfoRequested should be published \
             And Global state should be updated to RECEIVED.", async () => {
- 
+
      const bulkTransactionId = randomUUID();
      const bulkRequest: SDKSchemeAdapter.Outbound.V2_0_0.Types.bulkTransactionRequest = {
        bulkHomeTransactionID: "string",
@@ -163,22 +157,22 @@
      }
      const processSDKOutboundBulkRequestMessageObj = new ProcessSDKOutboundBulkRequestCmdEvt(sampleCommandEventData);
      await producer.sendCommandEvent(processSDKOutboundBulkRequestMessageObj);
- 
+
      await new Promise(resolve => setTimeout(resolve, messageTimeout));
      // Check the state in Redis
      console.log('bulk id: ', bulkTransactionId);
      const bulkState = await bulkTransactionEntityRepo.load(bulkTransactionId);
      expect(bulkState.state).toBe('RECEIVED');
- 
+
      // Check that the state of individual transfers in bulk to be RECEIVED
      const individualTransfers = await bulkTransactionEntityRepo.getAllIndividualTransferIds(bulkTransactionId);
      expect(individualTransfers.length).toBe(2);
      expect((await bulkTransactionEntityRepo.getIndividualTransfer(bulkTransactionId, individualTransfers[0])).state).toBe('RECEIVED');
      expect((await bulkTransactionEntityRepo.getIndividualTransfer(bulkTransactionId, individualTransfers[1])).state).toBe('RECEIVED');
- 
+
      // Check domain events published to kafka
      expect(domainEvents[0].getName()).toBe('SDKOutboundBulkPartyInfoRequestedDmEvt')
      // TODO Add asserts to check data contents of the domain event published to kafka
- 
+
    });
 });
