+++ modules/outbound-domain-event-handler/package.json
@@ -45,7 +45,7 @@
     },
     "devDependencies": {
         "@types/convict": "^6.1.1",
-        "@types/jest": "^29.0.2",
+        "@types/jest": "^29.0.3",
         "@types/node": "^18.7.18",
         "@types/node-cache": "^4.2.5",
         "@typescript-eslint/eslint-plugin": "^5.37.0",
@@ -53,7 +53,7 @@
         "eslint": "^8.23.1",
         "jest": "^29.0.3",
         "nodemon": "^2.0.19",
-        "npm-check-updates": "^16.1.2",
+        "npm-check-updates": "^16.1.3",
         "replace": "^1.2.1",
         "standard-version": "^9.5.0",
         "ts-jest": "^29.0.1",
