+++ modules/private-shared-lib/package.json
@@ -39,7 +39,7 @@
     "@types/node": "^18.7.18",
     "eslint": "^8.23.1",
     "jest": "^29.0.3",
-    "npm-check-updates": "^16.1.2",
+    "npm-check-updates": "^16.1.3",
     "replace": "^1.2.1",
     "standard-version": "^9.5.0",
     "ts-jest": "^29.0.1",
