+++ modules/private-shared-lib/src/events/outbound_domain_event/index.ts
@@ -38,3 +38,4 @@ export * from './sdk_outbound_bulk_quotes_request_processed';
 export * from './sdk_outbound_bulk_accept_quote_requested';
 export * from './sdk_outbound_bulk_accept_quote_received';
 export * from './sdk_outbound_bulk_response_sent';
+export * from './sdk_outbound_bulk_party_info_request_processed';
