+++ package.json
@@ -64,7 +64,7 @@
     "tslib": "^2.4.0"
   },
   "devDependencies": {
-    "@types/jest": "^29.0.2",
+    "@types/jest": "^29.0.3",
     "@types/node": "^18.7.18",
     "@types/node-cache": "^4.2.5",
     "@typescript-eslint/eslint-plugin": "^5.37.0",
@@ -76,7 +76,7 @@
     "husky": "^8.0.1",
     "jest": "^29.0.3",
     "nodemon": "^2.0.19",
-    "npm-check-updates": "^16.1.2",
+    "npm-check-updates": "^16.1.3",
     "replace": "^1.2.1",
     "standard-version": "^9.5.0",
     "ts-jest": "^29.0.1",
